update inflate value

+++ bioeye-assignment/ImageProcessor.h
@@ -14,11 +14,11 @@ private:
 	int detect_eyes(const cv::Mat&, std::vector<std::array<cv::Point, 2>>&);
 	std::array<cv::Point, 4> compute_eye_rect(const cv::Point, const cv::Point, const double, const double) const;
 	cv::Mat align_and_crop_eyes(const cv::Mat&, const std::array<cv::Point, 4>&) const;
-	void extract_aligned_eye_rect(const cv::Mat&, std::vector<cv::Mat>&, const std::array<cv::Point, 2>&, const float = 0.05f) const;
+	void extract_aligned_eye_rect(const cv::Mat&, std::vector<cv::Mat>&, const std::array<cv::Point, 2>&, const double = 0.05) const;
 
 
 	dlib::frontal_face_detector detector_;
 	dlib::shape_predictor shape_model_;
 	std::string landmark_model_path_ = "resources/shape_predictor_5_face_landmarks.dat";
-	float scale_factor = 0.4f;
+	double scale_factor = 0.4;
 };
